# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SunoSathi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me/onboarding:
    post:
      tags: [profile]
      operationId: completeOnboarding
      summary: Choose a role and complete onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingBody'
      responses:
        '200':
          description: Updated profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyProfile'
  /me/theme:
    post:
      tags: [profile]
      operationId: setThemePreference
      summary: Persist theme preference (light or dark)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThemeBody'
      responses:
        '200':
          description: Updated profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyProfile'
...
components:
...
  schemas:
...
    MyProfile:
      type: object
      required: [id, role, isAdmin, anonymousUsername, theme, hasOnboarded, createdAt]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        role:
          type: string
          enum: [user, listener]
        isAdmin:
          type: boolean
        anonymousUsername:
          type: string
        avatarSeed:
          type: string
        theme:
          type: string
          enum: [light, dark]
        hasOnboarded:
          type: boolean
        listenerProfile:
          oneOf:
            - $ref: '#/components/schemas/MyListenerProfile'
            - type: 'null'
        createdAt:
          type: string
          format: date-time
    OnboardingBody:
      type: object
      required: [role, anonymousUsername]
      properties:
        role:
          type: string
          enum: [user, listener]
        anonymousUsername:
          type: string
          minLength: 2
          maxLength: 24
        avatarSeed:
          type: string
        age:
          type: integer
          minimum: 13
          maximum: 80
    ThemeBody:
      type: object
      required: [theme]
      properties:
        theme:
          type: string
          enum: [light, dark]
...
    MyListenerProfile:
      type: object
      required: [id, userId, displayName, gender, bio, skills, photoUrl, applicationStatus, isOnline, ratingAverage, ratingCount]
      properties:
        id:
          type: string
        userId:
          type: string
        displayName:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        photoUrl:
          type: string
        applicationStatus:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: [string, 'null']
        isOnline:
          type: boolean
        lastSeenAt:
          type: [string, 'null']
          format: date-time
        ratingAverage:
          type: number
          format: float
        ratingCount:
          type: integer
...